perf: cache usd rate map from API
