+++ src/external-api.ts
@@ -12,38 +12,103 @@ import {
 } from "./interfaces";
 
 
-export const _getPoolsFromApi = memoize(
-    async (network: INetworkName, poolType: IPoolType, isLiteChain = false): Promise<IExtendedPoolDataFromApi> => {
-        const api = isLiteChain ? "https://api-core.curve.fi/v1/" : "https://api.curve.fi/api";
-        const url = `${api}/getPools/${network}/${poolType}`;
-        return await fetchData(url) ?? { poolData: [], tvl: 0, tvlAll: 0 };
-    },
-    {
-        promise: true,
-        maxAge: 5 * 60 * 1000, // 5m
-    }
-)
+const uncached_getPoolsFromApi = async (network: INetworkName, poolType: IPoolType, isLiteChain = false): Promise<IExtendedPoolDataFromApi> => {
+    const api = isLiteChain ? "https://api-core.curve.fi/v1/" : "https://api.curve.fi/api";
+    const url = `${api}/getPools/${network}/${poolType}`;
+    return await fetchData(url) ?? { poolData: [], tvl: 0, tvlAll: 0 };
+}
 
-export const _getAllPoolsFromApi = async (network: INetworkName, isLiteChain = false): Promise<IExtendedPoolDataFromApi[]> => {
-    if(isLiteChain) {
-        return await Promise.all([
-            _getPoolsFromApi(network, "factory-twocrypto", isLiteChain),
-            _getPoolsFromApi(network, "factory-tricrypto", isLiteChain),
-            _getPoolsFromApi(network, "factory-stable-ng", isLiteChain),
-        ]);
-    } else {
-        return await Promise.all([
-            _getPoolsFromApi(network, "main", isLiteChain),
-            _getPoolsFromApi(network, "crypto", isLiteChain),
-            _getPoolsFromApi(network, "factory", isLiteChain),
-            _getPoolsFromApi(network, "factory-crvusd", isLiteChain),
-            _getPoolsFromApi(network, "factory-eywa", isLiteChain),
-            _getPoolsFromApi(network, "factory-crypto", isLiteChain),
-            _getPoolsFromApi(network, "factory-twocrypto", isLiteChain),
-            _getPoolsFromApi(network, "factory-tricrypto", isLiteChain),
-            _getPoolsFromApi(network, "factory-stable-ng", isLiteChain),
-        ]);
+export const uncached_getAllPoolsFromApi = async (network: INetworkName, isLiteChain = false): Promise<Record<IPoolType, IExtendedPoolDataFromApi>> => {
+    const poolTypes = isLiteChain ? [
+        "factory-twocrypto",
+        "factory-tricrypto",
+        "factory-stable-ng",
+    ] as const : [
+        "main",
+        "crypto",
+        "factory",
+        "factory-crvusd",
+        "factory-eywa",
+        "factory-crypto",
+        "factory-twocrypto",
+        "factory-tricrypto",
+        "factory-stable-ng",
+    ] as const;
+    return Object.fromEntries(
+        await Promise.all(poolTypes.map(async (poolType) => {
+            const data = await uncached_getPoolsFromApi(network, poolType, isLiteChain);
+            return [poolType, data];
+        }))
+    )
+}
+
+export const uncached_getUsdPricesFromApi = async (allTypesExtendedPoolData:  IExtendedPoolDataFromApi[]): Promise<IDict<number>> => {
+    const priceDict: IDict<Record<string, number>[]> = {};
+    const priceDictByMaxTvl: IDict<number> = {};
+
+    for (const extendedPoolData of allTypesExtendedPoolData) {
+        for (const pool of extendedPoolData.poolData) {
+            const lpTokenAddress = pool.lpTokenAddress ?? pool.address;
+            const totalSupply = pool.totalSupply / (10 ** 18);
+            if(lpTokenAddress.toLowerCase() in priceDict) {
+                priceDict[lpTokenAddress.toLowerCase()].push({
+                    price: pool.usdTotal && totalSupply ? pool.usdTotal / totalSupply : 0,
+                    tvl: pool.usdTotal,
+                })
+            } else {
+                priceDict[lpTokenAddress.toLowerCase()] = []
+                priceDict[lpTokenAddress.toLowerCase()].push({
+                    price: pool.usdTotal && totalSupply ? pool.usdTotal / totalSupply : 0,
+                    tvl: pool.usdTotal,
+                })
+            }
+
+            for (const coin of pool.coins) {
+                if (typeof coin.usdPrice === "number") {
+                    if(coin.address.toLowerCase() in priceDict) {
+                        priceDict[coin.address.toLowerCase()].push({
+                            price: coin.usdPrice,
+                            tvl: pool.usdTotal,
+                        })
+                    } else {
+                        priceDict[coin.address.toLowerCase()] = []
+                        priceDict[coin.address.toLowerCase()].push({
+                            price: coin.usdPrice,
+                            tvl: pool.usdTotal,
+                        })
+                    }
+                }
+            }
+
+            for (const coin of pool.gaugeRewards ?? []) {
+                if (typeof coin.tokenPrice === "number") {
+                    if(coin.tokenAddress.toLowerCase() in priceDict) {
+                        priceDict[coin.tokenAddress.toLowerCase()].push({
+                            price: coin.tokenPrice,
+                            tvl: pool.usdTotal,
+                        });
+                    } else {
+                        priceDict[coin.tokenAddress.toLowerCase()] = []
+                        priceDict[coin.tokenAddress.toLowerCase()].push({
+                            price: coin.tokenPrice,
+                            tvl: pool.usdTotal,
+                        });
+                    }
+                }
+            }
+        }
     }
+
+    for(const address in priceDict) {
+        if (priceDict[address].length) {
+            const maxTvlItem = priceDict[address].reduce((prev, current) => +current.tvl > +prev.tvl ? current : prev);
+            priceDictByMaxTvl[address] = maxTvlItem.price
+        } else {
+            priceDictByMaxTvl[address] = 0
+        }
+    }
+
+    return priceDictByMaxTvl
 }
 
 export const _getSubgraphData = memoize(
